pyproject.toml
[project]
name = "devcoffee-schema"
version = "0.1.0"
description = "Devcoffee Schema Registry"
authors = [{ name = "Hieu Nguyen", email = "hieu.nguyen@devcoffee.tech" }]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

# =========================
# DEPENDENCIES
# =========================
dependencies = ["requests>=2.32.0", "django-environ>=0.12.1"]


# =========================
# DEV DEPENDENCIES
# =========================
[tool.pdm.dev-dependencies]
lint = ["ruff>=0.5.0"]


[tool.pdm]
distribution = true

# =========================
# RUFF
# =========================
[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = [".venv"]
src = ["src"]


[tool.pdm.scripts]
lint = "ruff check ."
fmt = "ruff format ."


[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # import sort
    "UP", # pyupgrade
    "B",  # bugbear
]
ignore = []


[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"


[tool.ruff.lint.isort]
known-first-party = ["devcoffee_schema"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
